Move struct Resolver out into its own file
---
